requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
chromadb
sentence-transformers
python-dotenv
pandas
fastapi
uvicorn
